fixed card layout and added joinClass function

+++ src/components/ClassCard.js
@@ -1,13 +1,15 @@
-import { deleteClass } from "../api/API";
+import { deleteClass, joinClass } from "../api/API";
+import { Heading } from "@aws-amplify/ui-react";
 
-import { TiDelete } from "react-icons/ti";
 import { CircleButton } from "./CircleButton";
 import { GiSoccerKick } from "react-icons/gi";
 import { AiOutlineTable } from "react-icons/ai";
-import { Heading } from "@aws-amplify/ui-react";
+import { BsSquare, BsCheckSquare } from "react-icons/bs";
 
 export const ClassCard = (props) => {
   const event = {
+    userId: props.userId,
+    isAdmin: props.isAdmin,
     id: props.id,
     type: props.type,
     day: props.day,
@@ -15,20 +17,26 @@ export const ClassCard = (props) => {
     minutes: props.minutes,
     duration: props.duration,
     spots: props.spots,
-    isAdmin: props.isAdmin,
-    //    confirmed: props.confirmed
+    confirmed: props.confirmed,
+  };
+
+  const handleJoinClass = async () => {
+    await joinClass(event.id, event.userId);
+    await props.refreshClasses();
   };
 
   const handleDeleteClass = async () => {
-    if (
-      window.confirm(
-        `Confirma a remoção da aula de ${event.type.toLowerCase()} que começaria às ${
-          event.hour
-        }:${event.minutes}`
-      )
-    ) {
-      await deleteClass({ id: event.id });
-      await props.refreshClasses();
+    if (event.isAdmin) {
+      if (
+        window.confirm(
+          `Confirma a remoção da aula de ${event.type.toLowerCase()} que começaria às ${
+            event.hour
+          }:${event.minutes}`
+        )
+      ) {
+        await deleteClass({ id: event.id });
+        await props.refreshClasses();
+      }
     }
   };
 
@@ -36,7 +44,12 @@ export const ClassCard = (props) => {
     <div style={styles.container}>
       <div style={styles.cardRow}>
         <div style={styles.circleButtonContainer}>
-          <CircleButton text={event.type} size="small" theme="dark">
+          <CircleButton
+            text={event.type}
+            size="small"
+            theme="dark"
+            onClick={() => handleDeleteClass()}
+          >
             {event.type == "Altinha" ? (
               <GiSoccerKick style={styles.icon} />
             ) : (
@@ -54,20 +67,19 @@ export const ClassCard = (props) => {
           </Heading>
         </div>
 
-        {event.isAdmin && (
-          <div
-            style={{
-              display: "flex",
-              flex: 0.5,
-              justifyContent: "center",
-            }}
-          >
-            <TiDelete
-              style={styles.iconDelete}
-              onClick={() => handleDeleteClass()}
+        <div style={styles.checkBoxContainer}>
+          {event.confirmed.includes(event.userId) ? (
+            <BsCheckSquare
+              onClick={() => handleJoinClass()}
+              style={styles.iconBlack}
             />
-          </div>
-        )}
+          ) : (
+            <BsSquare
+              onClick={() => handleJoinClass()}
+              style={styles.iconBlack}
+            />
+          )}
+        </div>
       </div>
     </div>
   );
@@ -89,40 +101,42 @@ const styles = {
     display: "flex",
     flex: 1,
     flexDirection: "row",
-    width: "100%",
+    width: "95%",
     justifyContent: "center",
     alignItems: "center",
-    margin: "1rem 0",
+    margin: "1rem",
   },
-  headerContainer: {
-    display: "flex",
-    flexDirection: "column",
-    flex: 1,
-    alignItems: "center",
-  },
-
   circleButtonContainer: {
     display: "flex",
     flex: 1,
     alignItems: "center",
     justifyContent: "center",
   },
-
+  headerContainer: {
+    display: "flex",
+    flexDirection: "column",
+    flex: 1.5,
+    alignItems: "center",
+  },
+  checkBoxContainer: {
+    display: "flex",
+    flex: 0.5,
+    alignItems: "center",
+    justifyContent: "space-around",
+  },
   headerText: {
     fontFamily: "Azonix",
     color: "#000",
     margin: "0.15rem",
   },
-
   icon: {
     fontSize: "2.5rem",
     cursor: "pointer",
     color: "#fff",
   },
-
-  iconDelete: {
+  iconBlack: {
     fontSize: "2rem",
     cursor: "pointer",
-    color: "#a00",
+    color: "#000",
   },
 };
